pyproject.toml
[project]
name = "pocketpaw"
version = "0.1.0"
description = "The AI agent that runs on your laptop, not a datacenter"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["ai", "agent", "telegram", "assistant", "automation"]
authors = [
    { name = "PocketPaw Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]

dependencies = [
    # Telegram Bot
    "python-telegram-bot>=21.0",
    # Web Server (QR Pairing)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "qrcode[pil]>=7.4",
    # System Tools
    "psutil>=5.9.0",
    "pyautogui>=0.9.54",
    "pillow>=10.0.0",
    # LLM Clients
    "httpx>=0.26.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    # Agent Backends
    "open-interpreter>=0.2.0",
    "claude-agent-sdk>=0.1.0",
    # Browser Automation
    "playwright>=1.40.0",
    # Config
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.22",
    # Scheduling
    "apscheduler>=3.10.0",
    "python-dateutil>=2.8.0",
    # Logging
    "rich>=13.0.0",
    "pytest-asyncio>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[project.scripts]
pocketpaw = "pocketclaw.__main__:main"

[project.urls]
Homepage = "https://github.com/pocketclaw/pocketclaw"
Repository = "https://github.com/pocketclaw/pocketclaw"
Issues = "https://github.com/pocketclaw/pocketclaw/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pocketclaw"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
